Clean up borrowing of guest VMSA structures
Borrowing of guest VMSA structures requires some sort of lifetime
checking to be robust.  The PerCpuShared object contains a spinlock to guard
against multiple accesses to the guest VMSA, but PerCpu::guest_vmsa
doesn't retain a reference to the spin lock to protect the access (the
spinlock reference goes out of scope with PerCpu::guest_vmsa returns).

This change move the guest_vmsa access to the GuestVmsaRef structure and
ties the VMSA lifetime to the reference lifetime (the LockGuard) so the lock
must remain held as long as the VMSA reference exists.  Furthermore,
because the VMSA reference is protected through the acquisition of a
spin lock, there is no reason to require a mutable PerCpu reference in
order to obtain a mutable VMSA reference, so this change also changes
references to the guest VMSA to use a shared reference to the current
CPU in order to reduce mutable borrow scopes.

Signed-off-by: Jon Lange <jlange@microsoft.com>