cpu/percpu: do not create a reference to uninitialized memory
In PerCpu::setup_ghcb(), a pointer to allocated but uninitialized
memory was being cast to a pointer to a GHCB struct. This pointer in
turn was used to call GHCB::init(), which takes &mut self, which
creates a reference to uninitialized memory. The GHCB struct consists
simply of scalar values, so it has no invalid representations, but
this is unnecessarily unsafe.

To avoid an unsafe block and potential future issues, allocate the
page zeroed, and make GHCB::init() take a VirtAddr which can properly
be initialized.

While we are at it, do not store the pointer in the PerCpu struct
until it has been properly initialized.

Signed-off-by: Carlos López <carlos.lopez@suse.com>