igvmbuilder: Move page directives to end an add option to sort by GPA
The order pages are defined in the IGVM file is generally not important.
However, in order to verify the output of the builder it is useful to
compare the output to the existing (working) C implementation of the
builder. To simplify this, this commit adds a sort option that ensures
all pages are described in ascending GPA order.

In addition, there is a requirement that directives defining parameters
are specified in the right order. After building the list of directives,
this commit unconditionally moves all Page directives to the end of the
list.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>